backend/requirements.txt
uvicorn
python-dotenv
fastapi
pytz
sqlalchemy
pymysql
requests
cryptography
fastapi-limiter
pyjwt